modified:   cadastrar_produto.py 	modified:   cadastrar_receita.py 	modified:   inicial_page.py 	modified:   mostrar_ingredientes.py 	modified:   pd.db

+++ cadastrar_produto.py
@@ -14,7 +14,7 @@ def create_database():
         CREATE TABLE IF NOT EXISTS estoque (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             produto TEXT NOT NULL UNIQUE,
-            quantidade INTEGER
+            quantidade REAL
         )
     ''')
     conn.commit()
